Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
praw==7.7.1
openai==1.3.5
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.3
requests==2.31.0
agents==0.1.0
aiofiles==23.2.1
httpx==0.25.2 